+++ include/vulnerable_website.h
@@ -1,6 +1,6 @@
 ﻿#pragma once
 
-#include <iostream>
-
-
+// http success
 const int SUCCESS_CODE = 200;
+
+
