+++ src/vulnerable_website.cpp
@@ -3,53 +3,14 @@ Author: Ashley Jones
 This small script logs into the website http://testphp.vulnweb.com.
 */
 
+// project include
+#include "utils.h"
 #include "vulnerable_website.h"
+// lib include
 #include "httplib.h"
+// STL include
 #include <iostream>
 #include <tuple> 
-
-
-void promptUserToStart() {
-	std::string line;
-	std::cout << "Please enter any key to start process: ";
-	std::getline(std::cin, line);
-	std::cout << std::endl;
-}
-
-
-void promptUserToEnd() {
-	std::string line;
-	std::cout << "Please enter any key to end process: ";
-	std::getline(std::cin, line);
-	std::cout << std::endl;
-}
-
-
-class LoginForm {
-	std::tuple<std::string, std::string, std::string> extractFormData(const std::string& data) {
-		// pretending as though we are parsing for now. 
-		return { "uname", "pass", "/userinfo.php" };
-	}
-
-public: 
-	std::string uname, pass, formScript;
-
-	LoginForm(const std::string& pageData) {
-		auto formData = extractFormData(pageData);
-		
-		this->uname = std::get<0>(formData);
-		this->pass = std::get<1>(formData);
-		this->formScript = std::get<2>(formData);
-	}
-
-	httplib::Params createLoginParams() {
-		httplib::Params loginParams{
-			  { uname, "test" },
-			  { pass, "test" }
-		};
-		return loginParams;
-	}
-};
 
 
 int main()
@@ -86,3 +47,4 @@ int main()
 
 	return 0;
 }
+
